move automl tests into subfolder
