+++ tests/regression_tests/automl/golden/adult_census_income.types.json
