+++ tests/regression_tests/automl/golden/mnist.types.json
